requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.0.0
tiktoken>=0.7.0
docker>=7.0.0
lancedb>=0.6.0
sentence-transformers>=2.7.0
pyarrow>=15.0.0
typer>=0.12.0
